[DeviceSanitizer] Move sanitizer predicates to new SanitizerUtils.h
This is follow-up of review comment
https://github.com/intel/llvm/pull/22901#discussion_r3752443031

isModuleUsingAsan/Msan/Tsan are sanitizer-specific, not generic SYCL
utilities, so move them out of SYCLLowerIR/SYCLUtils.{h,cpp} into a
new SanitizerUtils.{h,cpp}. Also add ASAN/MSAN/TSAN_KERNEL_METADATA_PREFIX
constants there so the "__{Asan,Msan,Tsan}KernelMetadata" prefix strings
have a single definition instead of being duplicated as literals across
SanitizerPostOptimizer.cpp, SanitizerPostSplitProcessing.cpp, and
clang's OffloadBundler.cpp.
